refactor(rust): extract transactions value types into submodule
Split the value types (TransitionKind, CommitMode, LogRetentionPolicy,
DoubletLink, Transition) and marker prefix constants out of
transactions/mod.rs into a new transactions/types.rs submodule so the
parent module stays under the 1000-line file-size gate enforced by
rust/scripts/check-file-size.rs.

mod.rs drops from 1002 to ~797 lines; public API is preserved via
re-exports.
